+++ src/main/java/edu/ie3/datamodel/io/csv/BufferedCsvWriter.java
@@ -27,10 +27,11 @@ public class BufferedCsvWriter extends BufferedWriter {
       "Direct appending is prohibited. Use write instead.";
 
   /**
-   * Build a new CsvBufferedWriter
+   * Build a new CsvBufferedWriter. The order of headline elements given in this constructor defines
+   * the order of columns in file
    *
    * @param filePath String representation of the full path to the target file
-   * @param headLineElements Elements of the csv head line
+   * @param headLineElements Elements of the csv headline
    * @param csvSep csv separator char
    * @param append true to append to an existing file, false to overwrite an existing file (if any),
    *     if no file exists, a new one will be created in both cases
@@ -47,7 +48,8 @@ public class BufferedCsvWriter extends BufferedWriter {
   /**
    * Build a new CsvBufferedWriter. This is a "convenience" Constructor. The absolute file path is
    * assembled by concatenation of {@code baseFolder} and {@code fileDefinition}'s file path
-   * information.
+   * information. The order of headline elements in {@code fileDefinition} defines the order of
+   * columns in file
    *
    * @param baseFolder Base folder, from where the file hierarchy should start
    * @param fileDefinition The foreseen shape of the file
@@ -73,7 +75,7 @@ public class BufferedCsvWriter extends BufferedWriter {
    */
   public synchronized void write(Map<String, String> entityFieldData)
       throws IOException, SinkException {
-    /* Check against eligible head line elements */
+    /* Check against eligible headline elements */
     if (entityFieldData.size() != headLineElements.length
         || !entityFieldData.keySet().containsAll(Arrays.asList(headLineElements)))
       throw new SinkException(
@@ -81,7 +83,8 @@ public class BufferedCsvWriter extends BufferedWriter {
               + String.join(",", headLineElements)
               + "'.");
 
-    String[] entries = entityFieldData.values().toArray(new String[0]);
+    String[] entries =
+        Arrays.stream(headLineElements).map(entityFieldData::get).toArray(String[]::new);
     writeOneLine(entries);
   }